Rewrite for 1.19.4

+++ common/src/main/java/net/heyimamethyst/fairyfactions/entities/FairyFishHookEntity.java
@@ -390,7 +390,7 @@ public class FairyFishHookEntity extends Projectile
                 double d0 = this.getX() + (double)(f1 * (float)this.timeUntilHooked * 0.1F);
                 double d1 = (double)((float)Mth.floor(this.getY()) + 1.0F);
                 double d2 = this.getZ() + (double)(f2 * (float)this.timeUntilHooked * 0.1F);
-                BlockState blockstate = serverlevel.getBlockState(new BlockPos(d0, d1 - 1.0D, d2));
+                BlockState blockstate = serverlevel.getBlockState(new BlockPos((int) d0, (int) (d1 - 1.0D), (int) d2));
                 if (serverlevel.getBlockState(new BlockPos((int)d0, (int)d1 - 1, (int)d2)).getMaterial() == net.minecraft.world.level.material.Material.WATER)
                 {
                     if (this.random.nextFloat() < 0.15F)
@@ -439,9 +439,9 @@ public class FairyFishHookEntity extends Projectile
                 double d4 = this.getX() + (double)(Mth.sin(f6) * f7) * 0.1D;
                 double d5 = (double)((float)Mth.floor(this.getY()) + 1.0F);
                 double d6 = this.getZ() + (double)(Mth.cos(f6) * f7) * 0.1D;
-                BlockState blockstate1 = serverlevel.getBlockState(new BlockPos(d4, d5 - 1.0D, d6));
+                BlockState blockstate1 = serverlevel.getBlockState(new BlockPos((int) d4, (int) (d5 - 1.0D), (int) d6));
 
-                if (serverlevel.getBlockState(new BlockPos(d4, d5 - 1.0D, d6)).getMaterial() == net.minecraft.world.level.material.Material.WATER)
+                if (serverlevel.getBlockState(new BlockPos((int) d4, (int) (d5 - 1.0D), (int) d6)).getMaterial() == net.minecraft.world.level.material.Material.WATER)
                 {
                     serverlevel.sendParticles(ParticleTypes.SPLASH, d4, d5, d6, 2 + this.random.nextInt(2), (double)0.1F, 0.0D, (double)0.1F, 0.0D);
                 }
